Run test scene from streaming assets
