+++ Assets/Scenes/manual.unity
@@ -330,7 +330,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: af9c7260f355e7b40877862842db6abb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  jsPath: Assets/Tests/TestFiles/cube_waves.js
+  jsFileName: cube_waves.js
 --- !u!4 &531854595
 Transform:
   m_ObjectHideFlags: 0
