+++ Assets/StreamingAssets.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a29b3f254fa1eee4cbb7c7b515130582
+guid: 4b844ec00db6fcc4e93159589f50b229
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
