--- Assets/Tests/TestFiles.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f56c4ccfb3955a041b90c5caffeebf6a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
